Silenced some warnings..
git-svn-id: https://svn.met.no/diana/trunk@2283 0ccb8871-b436-41a3-ae7d-e3ac59973202

+++ src/diVcrossField.cc
@@ -573,7 +573,7 @@ VcrossPlot* VcrossField::getCrossection(const miutil::miString& name,
       lastVcrossPlot->iundef = vcp->iundef;
 
       if(lastVcrossData.size()) {
-        for(int i=0;i<lastVcrossData.size();i++)
+        for(size_t i=0;i<lastVcrossData.size();i++)
           delete[] lastVcrossData[i];
       }
       lastVcrossData.clear();
